+++ frontend/openchat-agent/src/services/user/candid/idl.js
@@ -75,7 +75,6 @@ export const idlFactory = ({ IDL }) => {
   });
   const CommunityPermissions = IDL.Record({
     'create_public_channel' : CommunityPermissionRole,
-    'block_users' : CommunityPermissionRole,
     'update_details' : CommunityPermissionRole,
     'remove_members' : CommunityPermissionRole,
     'invite_users' : CommunityPermissionRole,